# magnifyBy2 faster