before:
`magnifyBy2 ........................ 14.762 ms     14.910 ms    ±0.153   x335`

after:
`magnifyBy2 ........................ 10.073 ms     10.267 ms    ±0.159   x486`